Intellipipe/mcp_server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
mcp>=0.1.0 # Or specific mcp[server] version
databricks-sdk>=0.20.0
pydantic>=2.0.0
pandas>=2.0.0
